+++ tests/test_checker/test_noqa.py
@@ -210,6 +210,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
 
     'WPS701': 2,
     'WPS702': 2,
+    'WPS703': 4,
 })